Make it possible to use the challenge-response API in discrete steps with public methods. (#6)
* Make it possible to use the challenge-response API in discrete steps, rather than the run() function with the callback, since the callback pattern will not always be the most convenient or suitable.
The ChallengeResponse::new_session() and ChallengeResponse::challenge_response() functions become public methods in the library.
The ChallengeResponseBuilder::build() function will no longer require the callback to be set. It remains optional. It is only required if ChallengeResponse::run() itself is invoked.
* Completely excise the evidence creation callback as a property of ChallengeResponse and ChallengeResponseBuilder.
Since it's only used in the context of ChallengeResponse::run(), it is simpler and more hygienic to pass it as a parameter to that function.
Also updated example code accordingly.

Signed-off-by: Paul Howard <paul.howard@arm.com>