Add Rust and C client support for the verification discovery API (#9)
Add Rust and C client support for the verification discovery API:

* First chunk of work for plumbing the endpoint discovery API through the Rust and C client layers.
The goal of this contribution is just to get the verification API data, concentrating on the public key for verification.
* Second chunk of work for Rust/C clients of discovery API.
Added the service-state and api-endpoints data to the gathered information
Added unit tests for the discovery API
Reworked the ChallengeResponse API so that it starts from a newSession API endpoint rather than from a base URL. This is because the correct idiom is to use the discovery API first, and then obtain the newSession endpoint from there, which then naturally pipes into the ChallengeResponseBuilder. This requires updating existing tests and example code.
* Plumb the media type list, version and service status fields from the Rust client through to the C wrapper.
The only thing left to plumb into the C wrapper is the map of endpoint names to relative URLs. Need to think about how to structure that.
* Plumb the API endpoint field through the C FFI layer. This now makes it functionally complete in both Rust and C.
* Update the C example program to call the discovery API before creating the session, and show how to use endpoint discovery to form the correct URL to create the session.
* Light-touch unit test (happy path) for FFI layer.
* Error handling pass over the FFI layer.
Re-uses the established convention of returning "stub" objects on failure, which contain largely empty values except for the error message.
Improved style by migrating most of the code to a safe helper function that returns a Rust-style result instead of a C result - allows for more succinct treatment of the several failure points, otherwise there would have to be a lot of ugly early returns.
* Add some missing docs.
* Address clippy faults.
* Improve for brevity.

Signed-off-by: Paul Howard <paul.howard@arm.com>